refactor(struct): 폴더 구조 변경
