+++ spec/auth/auth.mock.ts
@@ -3,11 +3,11 @@ import { JwtModule } from '@nestjs/jwt';
 
 import { mockMailService, mockMemberRepository, mockRedisRepository } from '../mock';
 
-import { AuthController } from '../../src/domain/auth/auth.controller';
-import { AuthService } from '../../src/domain/auth/auth.service';
+import { AuthController } from '../../src/domain/auth/controller';
+import { AuthService } from '../../src/domain/auth/service';
 
-import { Member } from '../../src/domain/member/member.schema';
-import { MemberRepository } from '../../src/domain/member/member.repository';
+import { Member } from '../../src/domain/member/schema';
+import { MemberRepository } from '../../src/domain/member/repository';
 
 import { RedisRepository } from '../../src/common/redis';
 import { MailService } from '../../src/common/utils/mail';
