+++ src/domain/activity/controller/index.ts
@@ -1,0 +1,1 @@
+export * from './activity.controller';
