+++ src/domain/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthController } from './controller';
+import { AuthService } from './service';
 
 import { MemberModule } from '../member/member.module';
