+++ src/domain/auth/dto/response/my-info.response.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 
-import { Role } from '../../../member/constant/Role';
-import { MyInfoLinks } from '../../../member/member.schema';
+import { Role } from '../../../member/constant';
+import { MyInfoLinks } from '../../../member/schema';
 
 export class MyInfoResponseDto {
   @ApiProperty({
