+++ src/domain/auth/service/index.ts
@@ -1,0 +1,1 @@
+export * from './auth.service';
