+++ src/domain/member/constant/index.ts
@@ -1,0 +1,1 @@
+export * from './Role';
