+++ src/domain/member/controller/index.ts
@@ -1,0 +1,2 @@
+export * from './member.controller';
+export * from './member.admin.controller';
