+++ src/domain/member/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './member.repository';
