+++ src/domain/member/service/index.ts
@@ -1,0 +1,2 @@
+export * from './member.service';
+export * from './member.admin.service';
